+++ packages/composites/proof-of-reserves/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @chainlink/proof-of-reserves-adapter
 
+## 1.14.0
+
+### Minor Changes
+
+- 2201988f2: add support for por_address_list protocol
+- 35fd235d5: Update Proof of Reserves to support Stader
+
+### Patch Changes
+
+- Updated dependencies [2201988f2]
+- Updated dependencies [35fd235d5]
+- Updated dependencies [35fd235d5]
+  - @chainlink/por-address-list-adapter@3.0.0
+  - @chainlink/eth-beacon-adapter@1.1.0
+  - @chainlink/stader-address-list-adapter@1.0.0
+  - @chainlink/por-indexer-adapter@1.2.20
+  - @chainlink/renvm-address-set-adapter@1.5.11
+  - @chainlink/wbtc-address-set-adapter@1.4.20
+
 ## 1.13.0
 
 ### Minor Changes